fix: reduce cron job for harvesting to every 15min
Initially I maid it run every hour, and configured all the jobs to run always, which means they will start and finish every hour.

But it looks like jobs are started but not finished, and only tagged as finished in the next cron job, which can be translated as the harvester runs every two hours.

We need also to understand how the jobs are added to the queue and processed. Some of the jobs were created, but not processed.